requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Limiter==3.5.0
requests==2.31.0
psycopg2-binary==2.9.7